+++ aiootp/ciphers/decipher_streams.py
@@ -319,10 +319,7 @@ class AsyncDecipherStream(CipherStreamProperties, metaclass=AsyncInit):
             raise Issue.invalid_length("data", len(data))
 
         async with ConcurrencyGuard(self._digesting_now):
-            if (
-                self._finalizing_now
-                and self._finalizing_now[0] not in self._digesting_now
-            ):
+            if await self._aconstant_time_final_context_is_done():
                 raise CipherStreamIsClosed
             data = io.BytesIO(data).read
             atest_block_id, append = self._buffer_shortcuts
@@ -607,10 +604,7 @@ class DecipherStream(CipherStreamProperties):
             raise Issue.invalid_length("data", len(data))
 
         with ConcurrencyGuard(self._digesting_now, probe_delay=0.0001):
-            if (
-                self._finalizing_now
-                and self._finalizing_now[0] not in self._digesting_now
-            ):
+            if self._constant_time_final_context_is_done():
                 raise CipherStreamIsClosed
             data = io.BytesIO(data).read
             atest_block_id, append = self._buffer_shortcuts